force socket.io to use websockets (browser)
This still allows connections using polling so that you can still
interact with the server using socket.io with the default settings.

This just saves a few HTTP requests and makes things a little cleaner.

(If the browser isn't new enough to support websockets then there would
be bigger problems to worry about!)