fix pathname discovery for socket events
Since 0a6424dde383414dcd2e7a2b1fd597917d256239 the referrer has not been
passed back to the server.
This is becasue that commit changed to only using websockets and
(socket.io at least) doesn't add the referrer info to the headers.

This commit changes to passing the pathname in a query string
when a socket connection is made.