improve redis backups
This changes the name of the backup entry.
The current state is now also stored alongside the log as well as
a list of the times that all of the backups were performed.

All values are in epoch time:

backup:log:xxxx           (new name for the log backups)
backup:stock_levels:xxxx  (newly stored state backup)
backup:all                (new list of all entries)