prioritise reading in beer information from Redis
This is the third (and final?) part of a series of commits that work on
the beer information. The others being:
  * dfb79e7735afe6e32a7c759c74405ac833fb6dd6
  * 4878f09b5ae378765a0548a245ca9b2ab30ff2ed

This adds the ability to save the beer information in the Redis database
then read it out on server start.

This means that if you upload beer information that differs from the
default, it will persist across restarts as it's stored in the DB.